Few architectural changes
