+++ dev.env
@@ -1,3 +1,5 @@
+HOST=0.0.0.0
+
 MONGO_URI=mongodb://database:27017
 MONGO_DBNAME=dev
 MONGO_COLLECTION=tasks
