+++ main.py
@@ -1,66 +1,17 @@
 import os
 import asyncio
-import json
-from typing import Any, Dict, Tuple
+from typing import Tuple
 
-from fastapi import FastAPI
-from fastapi.responses import HTMLResponse, JSONResponse
-from uvicorn import Config, Server
+from uvicorn import Server
 
 from amqp import MQConsumer, MQPublisher
 from robot_worker import RobotWorker
-from server_services import TaskService
-from store import MongoStore, BaseStore
-from models import Task
+from server import create_server
+from store import MongoStore
 
 
 def get_setting(name: str, prefix: str = ""):
     return os.getenv(f"{prefix}{name}")
-
-
-def create_server(
-    loop: asyncio.AbstractEventLoop,
-    store: BaseStore,
-    publisher: MQPublisher,
-    collection: str,
-) -> Server:
-    task_service = TaskService(
-        store=store, 
-        publisher=publisher, 
-        collection=collection
-    )
-
-    app = FastAPI()
-
-    @app.on_event("startup")
-    async def startup():
-        # Starting publisher connection
-        await task_service._publisher.start()
-
-    @app.get("/tasks")
-    async def task_form():
-        """ Getting tasks html form """
-        with open("templates/tasks/form.html", "r") as f:
-            return HTMLResponse(f.read())
-
-    @app.post("/tasks", status_code=201)
-    async def create_task(data: Dict[str, Any]):
-        document, is_created = await task_service.create_task(data)
-
-        if is_created:
-            return {"message": f"task {data} is queued"} 
-        else:
-            return JSONResponse(
-                status_code=200, 
-                content={"message": json.loads(Task(**document).json())}
-            )
-
-    config = Config(app=app, loop=loop, host="0.0.0.0")
-    return Server(config)
-
-
-def create_robot(store: BaseStore, consumer: MQConsumer, collection: str) -> RobotWorker:
-    return RobotWorker(store=store, consumer=consumer, collection=collection)
 
 
 def create_app(
@@ -88,9 +39,10 @@ def create_app(
         loop=loop, 
         store=store, 
         publisher=publisher, 
-        collection=collection
+        collection=collection,
+        host=get_setting("HOST", prefix)
     )
-    robot = create_robot(store=store, consumer=consumer, collection=collection)
+    robot = RobotWorker(store=store, consumer=consumer, collection=collection)
 
     return server, robot
