+++ test.env
@@ -1,3 +1,5 @@
+TEST_HOST=0.0.0.0
+
 TEST_MONGO_URI=mongodb://database:27017
 TEST_MONGO_DBNAME=test_db
 TEST_MONGO_COLLECTION=tasks
